move dragElement into utils
