+++ src/components/note.js
@@ -1,5 +1,5 @@
 import "./note.css";
-import { moveToTop, noteGallery } from "../utils";
+import { moveToTop, noteGallery, dragElement } from "../utils";
 import { saveSession } from "../session";
 
 window.addEventListener("beforeunload", saveSession);
@@ -28,36 +28,6 @@ export function createNote(x, y, offset = true) {
   note.addEventListener("mousedown", dragElement(note));
   note.setAttribute("draggable", "true");
   return note;
-}
-
-function dragElement(elmnt) {
-  let cursorX, cursorY, cursorXDelta, cursorYDelta;
-  elmnt.onmousedown = dragStart;
-
-  function dragStart(e) {
-    e.preventDefault();
-    cursorX = e.clientX;
-    cursorY = e.clientY;
-    // capture dragging movements outside of elmnt for smoothness
-    document.onmouseup = dragEnd;
-    document.onmousemove = moveElement;
-  }
-
-  function moveElement(e) {
-    e.preventDefault();
-    cursorXDelta = cursorX - e.clientX;
-    cursorYDelta = cursorY - e.clientY;
-    cursorX = e.clientX;
-    cursorY = e.clientY;
-    // update element's position
-    elmnt.style.left = elmnt.offsetLeft - cursorXDelta + "px";
-    elmnt.style.top = elmnt.offsetTop - cursorYDelta + "px";
-  }
-
-  function dragEnd() {
-    document.onmouseup = null;
-    document.onmousemove = null;
-  }
 }
 
 export function addNote(x, y) {
